fix: x264 float64→float32 cast error, resolve README merge conflict, update version to 0.0.5

+++ Goylord-Client/cmd/agent/capture/h264_encoder_x264.go
@@ -104,7 +104,7 @@ func ensureH264EncoderLocked(width, height int) error {
 		Preset:       "veryfast",
 		Profile:      "main",
 		RateControl:  "crf",
-		RateConstant: float64(targetH264CRF(width, height, fps)),
+		RateConstant: targetH264CRF(width, height, fps),
 		LogLevel:     x264.LogError,
 	}
 
@@ -201,7 +201,7 @@ func ensurebackstageH264EncoderLocked(width, height int) error {
 		Preset:       "veryfast",
 		Profile:      "main",
 		RateControl:  "crf",
-		RateConstant: float64(targetH264CRF(width, height, fps)),
+		RateConstant: targetH264CRF(width, height, fps),
 		LogLevel:     x264.LogError,
 	}
 	enc, err := x264.NewEncoder(&backstageH264Buf, opts)
@@ -280,7 +280,7 @@ func ensureWebcamH264EncoderLocked(width, height int) error {
 		Preset:       "veryfast",
 		Profile:      "main",
 		RateControl:  "crf",
-		RateConstant: float64(targetH264CRF(width, height, fps)),
+		RateConstant: targetH264CRF(width, height, fps),
 		LogLevel:     x264.LogError,
 	}
 	enc, err := x264.NewEncoder(&webcamH264Buf, opts)
